# resources/api/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Arkon API",
        "version": "1.0.0",
        "description": "Arkon's public developer API. Guide: docs/api.md. Anonymous requests read published content; a personal access token (Authorization: Bearer) adds the scopes it was given, never more than its user's role allows. Errors: {\"error\": {\"code\", \"message\", \"fields\"?}}."
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "The site the API is reached at"
        }
    ],
    "security": [
        [],
        {
            "bearerAuth": []
        }
    ],
...
    "paths": {
...
        "/media": {
            "get": {
                "operationId": "listMedia",
                "tags": [
                    "Media"
                ],
                "summary": "Images live pages use (anonymous) or the whole library (read:media)",
                "security": [
                    [],
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "description": "Title, file name or alt text.",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "mime_type",
                        "in": "query",
                        "required": false,
                        "description": "e.g. image/jpeg.",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "description": "read:media only.",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "library",
                                "trash"
                            ]
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "required": false,
                        "description": "Default -uploaded_at.",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "uploaded_at",
                                "-uploaded_at",
                                "title",
                                "-title",
                                "filesize",
                                "-filesize"
                            ]
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "description": "Page number, from 1.",
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "description": "Items per page (default 10, at most 100).",
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "maximum": 100
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "required": [
                                        "data",
                                        "meta",
                                        "links"
                                    ],
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Media"
                                            }
                                        },
                                        "meta": {
                                            "$ref": "#/components/schemas/PaginationMeta"
                                        },
                                        "links": {
                                            "$ref": "#/components/schemas/PaginationLinks"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/Error"
                    },
                    "401": {
                        "$ref": "#/components/responses/Error"
                    },
                    "403": {
                        "$ref": "#/components/responses/Error"
                    }
                }
            },
            "post": {
                "operationId": "uploadMedia",
                "tags": [
                    "Media"
                ],
                "summary": "Upload an image (the same pipeline as the admin: validated, original kept, WebP sizes)",
                "security": [
                    {
                        "bearerAuth": [
                            "write:media"
                        ]
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "required": [
                                    "file"
                                ],
                                "properties": {
                                    "file": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "title": {
                                        "type": "string"
                                    },
                                    "alt": {
                                        "type": "string"
                                    },
                                    "caption": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "required": [
                                        "data"
                                    ],
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Media"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Error"
                    },
                    "403": {
                        "$ref": "#/components/responses/Error"
                    },
                    "413": {
                        "$ref": "#/components/responses/Error"
                    },
                    "422": {
                        "$ref": "#/components/responses/Error"
                    }
                }
            }
        },
        "/media/{id}": {
            "get": {
                "operationId": "getMedia",
                "tags": [
                    "Media"
                ],
                "summary": "One image",
                "security": [
                    [],
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "The id.",
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "required": [
                                        "data"
                                    ],
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Media"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/Error"
                    }
                }
            },
            "patch": {
                "operationId": "updateMedia",
                "tags": [
                    "Media"
                ],
                "summary": "Change title, alt text, caption or description",
                "security": [
                    {
                        "bearerAuth": [
                            "write:media"
                        ]
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "The id.",
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MediaWrite"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "required": [
                                        "data"
                                    ],
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Media"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Error"
                    },
                    "403": {
                        "$ref": "#/components/responses/Error"
                    },
                    "404": {
                        "$ref": "#/components/responses/Error"
                    },
                    "409": {
                        "$ref": "#/components/responses/Error"
                    },
                    "422": {
                        "$ref": "#/components/responses/Error"
                    }
                }
            },
            "delete": {
                "operationId": "trashMedia",
                "tags": [
                    "Media"
                ],
                "summary": "Move to the Trash (pages that use it keep working)",
                "security": [
                    {
                        "bearerAuth": [
                            "write:media"
                        ]
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "The id.",
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Done"
                    },
                    "401": {
                        "$ref": "#/components/responses/Error"
                    },
                    "403": {
                        "$ref": "#/components/responses/Error"
                    },
                    "404": {
                        "$ref": "#/components/responses/Error"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "description": "A personal access token (arkon_pat_…). Scopes: read:pages, write:pages, read:posts, write:posts, write:categories, write:tags, read:media, write:media, read:forms, read:form_entries, read:users."
            }
        },
        "responses": {
            "Error": {
                "description": "An error",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Error"
                        }
                    }
                }
            }
        },
        "schemas": {
            "Error": {
                "type": "object",
                "required": [
                    "error"
                ],
                "properties": {
                    "error": {
                        "type": "object",
                        "required": [
                            "code",
                            "message"
                        ],
                        "properties": {
                            "code": {
                                "type": "string"
                            },
                            "message": {
                                "type": "string"
                            },
                            "fields": {
                                "type": "object",
                                "additionalProperties": {
                                    "type": "array",
                                    "items": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "PaginationMeta": {
                "type": "object",
                "required": [
                    "page",
                    "per_page",
                    "total",
                    "total_pages"
                ],
                "properties": {
                    "page": {
                        "type": "integer"
                    },
                    "per_page": {
                        "type": "integer"
                    },
                    "total": {
                        "type": "integer"
                    },
                    "total_pages": {
                        "type": "integer"
                    }
                }
            },
            "PaginationLinks": {
                "type": "object",
                "required": [
                    "self",
                    "first",
                    "prev",
                    "next",
                    "last"
                ],
                "properties": {
                    "self": {
                        "type": "string"
                    },
                    "first": {
                        "type": "string"
                    },
                    "prev": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "next": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "last": {
                        "type": "string"
                    }
                }
            },
...
            "MediaSize": {
                "type": "object",
                "required": [
                    "url",
                    "width",
                    "height",
                    "mime_type",
                    "filesize"
                ],
                "properties": {
                    "url": {
                        "type": "string"
                    },
                    "width": {
                        "type": "integer"
                    },
                    "height": {
                        "type": "integer"
                    },
                    "mime_type": {
                        "type": "string"
                    },
                    "filesize": {
                        "type": "integer"
                    }
                }
            },
            "Media": {
                "type": "object",
                "required": [
                    "id",
                    "title",
                    "filename",
                    "mime_type",
                    "width",
                    "height",
                    "filesize",
                    "alt",
                    "caption",
                    "url",
                    "thumbnail",
                    "variants",
                    "formats",
                    "uploaded_at"
                ],
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "title": {
                        "type": "string"
                    },
                    "filename": {
                        "type": "string"
                    },
                    "mime_type": {
                        "type": "string"
                    },
                    "width": {
                        "type": "integer"
                    },
                    "height": {
                        "type": "integer"
                    },
                    "filesize": {
                        "type": "integer"
                    },
                    "alt": {
                        "type": "string"
                    },
                    "caption": {
                        "type": "string"
                    },
                    "url": {
                        "type": "string"
                    },
                    "thumbnail": {
                        "type": [
                            "object",
                            "null"
                        ],
                        "properties": {
                            "url": {
                                "type": "string"
                            },
                            "width": {
                                "type": "integer"
                            },
                            "height": {
                                "type": "integer"
                            }
                        }
                    },
                    "variants": {
                        "type": "object",
                        "description": "Responsive WebP sizes keyed by width.",
                        "additionalProperties": {
                            "$ref": "#/components/schemas/MediaSize"
                        }
                    },
                    "formats": {
                        "type": "object",
                        "required": [
                            "webp",
                            "avif"
                        ],
                        "properties": {
                            "webp": {
                                "type": "boolean"
                            },
                            "avif": {
                                "type": "boolean"
                            }
                        }
                    },
                    "uploaded_at": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "format": "date-time"
                    },
                    "public": {
                        "type": "boolean",
                        "description": "read:media only."
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "library",
                            "trash"
                        ],
                        "description": "read:media only."
                    },
                    "description": {
                        "type": "string",
                        "description": "read:media only."
                    },
                    "version": {
                        "type": "integer",
                        "description": "read:media only."
                    }
                }
            },
            "MediaWrite": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                    "title": {
                        "type": "string"
                    },
                    "alt": {
                        "type": "string"
                    },
                    "caption": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "version": {
                        "type": "integer",
                        "description": "The version you read; a newer one is a 409."
                    }
                }
            },
...
        }
    }
}